+++ stdlib/REPL/src/REPL.jl
@@ -207,7 +207,7 @@ function run_frontend(repl::BasicREPL, backend::REPLBackendRef)
         interrupted = false
         while true
             try
-                line *= readline(repl.terminal, chomp=false)
+                line *= readline(repl.terminal, keep=true)
             catch e
                 if isa(e,InterruptException)
                     try # raise the debugger if present
@@ -378,7 +378,7 @@ An editor may have converted tabs to spaces at line """
 
 function hist_getline(file)
     while !eof(file)
-        line = readline(file, chomp=false)
+        line = readline(file, keep=true)
         isempty(line) && return line
         line[1] in "\r\n" || return line
     end
@@ -979,10 +979,10 @@ function setup_interface(
                 @goto writeback
             end
             Base.edit(linfos[n][1], linfos[n][2])
-            Base.LineEdit.refresh_line(s)
+            LineEdit.refresh_line(s)
             return
             @label writeback
-            write(Base.LineEdit.buffer(s), str)
+            write(LineEdit.buffer(s), str)
             return
         end,
     )
@@ -1112,7 +1112,7 @@ function run_frontend(repl::StreamREPL, backend::REPLBackendRef)
         if have_color
             print(repl.stream, input_color(repl))
         end
-        line = readline(repl.stream, chomp=false)
+        line = readline(repl.stream, keep=true)
         if !isempty(line)
             ast = Base.parse_input_line(line)
             if have_color